requirements.txt
django-heroku
gunicorn
asgiref
psycopg2
whitenoise